requirements.txt
requests==2.31.0
pytest==7.4.0
allure-pytest==2.13.2
python-dotenv==1.0.0
faker==19.6.2